Only write version information on real robot (#1510)
Writing to the versions file throws an exception on Windows, and might
write weird files on Linux.

+++ wpilibj/src/main/java/edu/wpi/first/wpilibj/RobotBase.java
@@ -237,23 +237,25 @@ public abstract class RobotBase implements AutoCloseable {
       return;
     }
 
-    try {
-      final File file = new File("/tmp/frc_versions/FRC_Lib_Version.ini");
+    if (isReal()) {
+      try {
+        final File file = new File("/tmp/frc_versions/FRC_Lib_Version.ini");
 
-      if (file.exists()) {
-        file.delete();
+        if (file.exists()) {
+          file.delete();
+        }
+
+        file.createNewFile();
+
+        try (OutputStream output = Files.newOutputStream(file.toPath())) {
+          output.write("Java ".getBytes(StandardCharsets.UTF_8));
+          output.write(WPILibVersion.Version.getBytes(StandardCharsets.UTF_8));
+        }
+
+      } catch (IOException ex) {
+        DriverStation.reportError("Could not write FRC_Lib_Version.ini: " + ex.toString(),
+                ex.getStackTrace());
       }
-
-      file.createNewFile();
-
-      try (OutputStream output = Files.newOutputStream(file.toPath())) {
-        output.write("Java ".getBytes(StandardCharsets.UTF_8));
-        output.write(WPILibVersion.Version.getBytes(StandardCharsets.UTF_8));
-      }
-
-    } catch (IOException ex) {
-      DriverStation.reportError("Could not write FRC_Lib_Version.ini: " + ex.toString(),
-          ex.getStackTrace());
     }
 
     boolean errorOnExit = false;
